Reactor Management
New save technology utilizes tables to save data in just a fraction of
the coding, now all settings can be stored in one place by just creating
them on the fly and reloading them anytime, anywhere.

Monitor support allows monitors to be placed in a wired network or in
direct contact with the computer, thus allowing the rectors status to be
displayed on the big screen.